[monitoring/uss_qualifier] Move utils files to rid package
